Added search by Genre, updated the SearchBar styling, Moved RenderMedia to Components from the pages folder
