+++ src/Components/Search.module.css
@@ -1,60 +1,30 @@
-.flexContainer {
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: space-around;
-    align-items: center;
-}
-.flexPadding {
-    padding: 10px;
-}
 .container {
     display: flex;
     align-items: center;
     background-color: white;
-    border-radius: 3rem;
+    border-radius: 1.5rem;
     box-shadow: 0 0.063rem 0.375rem rgba(32 33 36 0.28);
     padding: 0.5rem;
-    width: 28rem;
-    height: 6rem;
+    width: 20rem;
     margin: 0 auto;
 }
 .input {
     flex: 0.063rem;
     border: none;
     outline: none;
-    width: 18rem;
-    font-size: 20px;
-    font-weight: lighter;
-    padding: 0.5rem 0rem 0.5rem 0.5rem;
+    font-size: 1rem;
+    font-weight: 31.25rem;
+    padding: 0.5rem;
 }
 .button {
     border: none;
     outline: none;
     background-color: #fecb00;
-    border-radius: 3rem;
+    border-radius: 1.5rem;
     padding: 0.5rem;
-    margin-left: 2rem;
-    margin-right: 0.5rem;
-    width: 5rem;
-    height: 5rem;
+    margin-left: 0.5rem;
     cursor: pointer;
 }
 .logo {
     background-color: white;
-}
-.sort {
-    font-size: 20px;
-    font-weight: lighter;
-    text-align: center;
-    width: 18.75rem;
-    height: 6rem;
-    padding: 10px 20px 10px 10px;
-}
-.dropdown {
-    font-size: 18px;
-    width: 19.25rem;
-    font-weight: lighter;
-    padding: 10px 8px 10px 10px;
-    text-align: center;
-    border-radius: 2rem;
 }
